renderer: Move terrain and world rendering into subdirectories.
