+++ libopenage/renderer/stages/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(terrain/)
+add_subdirectory(world/)
