+++ libopenage/renderer/stages/terrain/terrain_mesh.h
