+++ libopenage/renderer/stages/terrain/terrain_model.h
