+++ libopenage/renderer/stages/terrain/terrain_render_entity.h
