+++ libopenage/renderer/stages/terrain/terrain_renderer.h
