+++ libopenage/renderer/stages/world/CMakeLists.txt
