+++ libopenage/renderer/stages/world/world_render_entity.h
