+++ libopenage/renderer/stages/world/world_renderer.cpp
