+++ libopenage/renderer/stages/world/world_renderer.h
